requirements.txt
asgiref==3.8.1
Django==5.0.2
djangorestframework==3.14.0
psycopg2-binary==2.9.9
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
social-auth-app-django==5.4.0
